change response of eslint-bridge to include ucfg paths
