+++ eslint-bridge/src/linting/eslint/linter/issues/transform.ts
@@ -22,11 +22,43 @@ import { Linter, Rule, SourceCode } from 'eslint';
 import { decodeSonarRuntime } from './decode';
 import { Issue } from './issue';
 import { convertMessage } from './message';
+import { extractCognitiveComplexity, extractHighlightedSymbols } from './extract';
+import { SymbolHighlight } from '../visitors';
+
+/**
+ * The result of linting a source code
+ *
+ * ESLint API returns what it calls messages as results of linting a file.
+ * A linting result in the context of the analyzer includes more than that
+ * as it needs not only to transform ESLint messages into SonarQube issues
+ * as well as analysis data about the analyzed source code, namely symbol
+ * highlighting and cognitive complexity.
+ *
+ * @param issues the issues found in the code
+ * @param highlightedSymbols the symbol highlighting of the code
+ * @param cognitiveComplexity the cognitive complexity of the code
+ */
+export type LintingResult = {
+  issues: Issue[];
+  ucfgPaths: string[];
+  highlightedSymbols: SymbolHighlight[];
+  cognitiveComplexity?: number;
+};
 
 /**
  * Transforms ESLint messages into SonarQube issues
  *
+ * The result of linting a source code requires post-linting transformations
+ * to return SonarQube issues. These transformations include extracting ucfg
+ * paths, decoding issues with secondary locations as well as converting
+ * quick fixes.
+ *
+ * Besides issues, a few metrics are computing during linting in the form of
+ * an internal custom rule execution, namely cognitive complexity and symbol
+ * highlighting. These custom rules also produce issues that are extracted.
+ *
  * Transforming an ESLint message into a SonarQube issue implies:
+ * - extracting UCFG rule file paths
  * - converting ESLint messages into SonarQube issues
  * - converting ESLint fixes into SonarLint quick fixes
  * - decoding encoded secondary locations
@@ -34,17 +66,35 @@ import { convertMessage } from './message';
  *
  * @param messages ESLint messages to transform
  * @param ctx contextual information
- * @returns the transformed issues
+ * @returns the linting result
  */
 export function transformMessages(
   messages: Linter.LintMessage[],
   ctx: { sourceCode: SourceCode; rules: Map<string, Rule.RuleModule> },
-): Issue[] {
-  return messages
-    .map(message => convertMessage(ctx.sourceCode, message))
-    .map(issue => (issue ? decodeSonarRuntime(ctx.rules.get(issue.ruleId), issue) : null))
-    .filter((issue): issue is Issue => issue !== null)
-    .map(normalizeLocation);
+): LintingResult {
+  const issues: Issue[] = [];
+  const ucfgPaths: string[] = [];
+
+  for (const message of messages) {
+    if (message.ruleId === 'ucfg') {
+      ucfgPaths.push(message.message);
+    } else {
+      let issue = convertMessage(ctx.sourceCode, message);
+      if (issue !== null) {
+        issue = normalizeLocation(decodeSonarRuntime(ctx.rules.get(issue.ruleId), issue));
+        issues.push(issue);
+      }
+    }
+  }
+
+  const highlightedSymbols = extractHighlightedSymbols(issues);
+  const cognitiveComplexity = extractCognitiveComplexity(issues);
+  return {
+    issues,
+    ucfgPaths,
+    highlightedSymbols,
+    cognitiveComplexity,
+  };
 }
 
 /**
